==> rmd/01_create_samples.Rmd <==
---
title: "Create samples"
output:
  html_document:
    df_print: paged
---


```{r}
library(pacman)
p_load(dplyr, tidyverse, DBI, bigrquery, feather, pander, bigrquery, glue)
```


```{r}
con <- dbConnect(
  bigrquery::bigquery(),
  project = "dataforseo-bigquery",
  billing = "dataforseo-bigquery"
)

tbl1 <- tbl(con, "dataforseo-bigquery.dataforseo_data.keyword_data")
```

```{r}
write_keyword_files <- function(min, max, n, n_files){
  sql <- glue("SELECT * FROM `dataforseo-bigquery.dataforseo_data.keyword_data` 
          WHERE location = 2840 
          AND keyword_info_search_volume >= {min}
          AND keyword_info_search_volume < {max}
          ORDER BY RAND()
          LIMIT {n}")
  tb <- bq_project_query("dataforseo-bigquery", sql)
  df <- bq_table_download(tb, max_results = as.numeric(n)) %>% 
    mutate(g = (row_number() -1) %/% (n() / n_files))
  
  for (i in 0:(n_files-1)){
    write_csv(df %>% filter(g == i) %>% select(keyword), 
              glue("keywords_{min}_to_{max}/{i}.csv"), col_names = F)
  }  
}

write_keyword_files("100", "500", "1000000", 100)
write_keyword_files("500", "1000", "1000000", 100)
write_keyword_files("1000", "10000", "1000000", 100)
write_keyword_files("10000", "1000000000", "1000000", 100)
```
