==> rmd/03_final_1911.Rmd <==
---
title: "We Analyzed 306M Keywords. Here's What We Learned About Google Searches"
author: 
- "Lead: Brian Dean ([backlinko.com](https://backlinko.com/))"
- "Support: FrontPage Data ([frontpagedata.com](https://frontpagedata.com/))"
date: "Last updated on `r Sys.Date()`"
output:
  html_document:
    theme: paper
    highlight: kate
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: true
editor_options:
  chunk_output_type: inline
---

```{=html}
<style>
.list-group-item.active, .list-group-item.active:hover, .list-group-item.active:focus {
  background-color: #00d188;
  border-color: #00d188;
}

body {
  font-family: montserrat;
  color: #444444;
  font-size: 14px;
}

h1 {
  font-weight: bold;
  font-size: 28px;
}

h1.title {
  font-size: 30px;
  color: #00d188;
}

h2 {
  font-size: 24px;
}

h3 {
  font-size: 18px;
}

.main-container {
  max-width: 1800px;
  margin-left: auto;
  margin-right: auto;
}
</style>
```
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE, fig.showtext = TRUE, out.width = "100%", out.height = "100%")
knitr::knit_hooks$set(inline = function(x) {
  prettyNum(x, big.mark = ",", small.mark = ",", scientific = F)
})
#plots <- readRDS(here::here("proc_data/plots.RDS"))
include_png <- function(item) {
  knitr::include_graphics(
    here::here("plots", "png", paste0(make.names(item), ".png")))
  } 
```

```{r}
library(tidyverse)
library(gt)
```

<br>

# 1 Introduction

Keyword data informs and impacts every other SEO task, including finding content topics, on-page SEO, email outreach, and content promotion. While a wealth of companies provide granular keyword data, a birds-eye view of Google Searches is missing in the industry. This large-scale analysis aims to fill the gap.

# 2 Methodology

As a data partner, DataForSEO provided the raw data for the analysis through their keyword database [<https://dataforseo.com/databases>]. The analysis was limited to US keyword data that accounted for around 306 keyword searches. Keyword monthly searchers with volume of 0 were removed from the data set. The data was queried from a Google BigQuery database.

To attain additional insights from the data, we enriched a random sample of around 2.5 million keywords with the Ahrefs Keyword Explorer. This includes the data on difficulty, return rate, clicks, region volume, and SERP features.

Below an overview of the data:

```{r overview}
tbl <- tribble(~Statistic, ~Value,
  "Total number of searches", "~306 million",
  "Total volume of searches", "~32 billion",
  "Searches with missing volume", "0.51%",
  "Mean search volume", "989",
  "Median search volume", "10",
  "Mean CPC", "0.61")

knitr::kable(tbl, caption = "Overview")
```

(For how the volume was calculated, see appendix.)

<br>

# 3 Research Findings

<br>

## 3.1 Top searches

These are the most popular searches, with volume based on *ahref*:

```{r top, echo=FALSE}
title_description_table <- read.csv(here::here("plots", "csv", "top50_table.csv")) %>% 
    mutate(Keyword = keyword, Volume = scales::percent(volume, accuracy = 0.001)) %>% 
    select(Keyword, Volume)
knitr::kable(title_description_table, caption = "Searches with highest volume")
```

<br> <br>

## 3.2 Search volume

The search volume is concentrated among the top searches:

```{r high_volume}
title_description_table <- read.csv(here::here("plots", "csv", "top_volume.csv")) %>%
    mutate(volume = scales::percent(volume, accuracy = 0.1)) %>% 
    rename(Searches = top, Volume = volume)
knitr::kable(title_description_table, caption = "High volume among top searches")
```

If we divide the searches into categories based on how much volume they receive, this gives an overview:

```{r number_searches, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "proportion_searches.png"))
```

```{r volume_searches, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "proportion_search_volume.png"))
```

These figures together show that low volume searches (individual searches with volume less than 100), constitute 91.8% of all searches, but only 3.3% of all volume. Or stated in another way, 91.8% of keywords get 1-100 volume a month. Whereas very high volume searches (individual searches with volume above 10,000) constitute only 0.2% of all searches, but have 75.2% of all volume.

The top searches have high volume compared to the rest, and the search volume level quickly levels off. The picture looks slightly different depending on whether we use the values from *DataforSEO*:

```{r}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_by_searches_dataforseo.png"))
```

Or *ahref*:

```{r}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_by_searches_ahref.png"))
```

<br> <br>

## 3.3 Spell types

If a misspelling is recognized, a so-called spell type is suggested. There are three types of spell shown in the table below. \~1.4% of searches have a spell type, but those that do have tend to have high volume.

```{r spell_types_count, echo=FALSE}
knitr::include_graphics(here::here("plots", "0911", "spell_types.png"))
```

<br> <br>

## 3.4 Questions in searches

14.1% of searches are in the form of a question. "how" is the most common question word.

```{r questions_count, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "questions_count.png"))
```

<br> <br>

## 3.5 Stopwords

A list of stopwords was qcquired from here: <https://meta.wikimedia.org/wiki/Stop_word_list/google_stop_word_list#English>

"how" and "the" are the most common stopwords, which are present in 6-8% of searches.

```{r stopwords_count, echo=FALSE}
knitr::include_graphics(here::here("plots", "0911", "stopwords_cloud_dark.png"))
```

*A colorful version:*

```{r stopwords_count_colorful, echo=FALSE}
knitr::include_graphics(here::here("plots", "0911", "stopwords_cloud_colors.png"))
```

<br> <br>

## 3.6. Length of search term

The most searched queries have length 5-10 characters, and falls continuously for search queries longer or shorter than that.

```{r keyword_length, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_keywordlength.png"))
```

The mean keyword length in characters is 11.2, and the median is 8.5.

```{r keyword_words_volume, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_words_searchterm.png"))
```

Most searches consist of a single word, with volume quickly falling as the number of words increase.

Search terms that are 1-3 words in length get \~10x as much volume on average as search terms that are 5+ words in length.

<br> <br>

## 3.7 Keyword info categories

Internet & Telecom is the keyword category with the highest mean volume.

```{r volume_mean_category_ordered, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_mean_category_ordered.png"))
```

([data table](https://github.com/danielkupka/2020-google-searches/blob/master/proc_data/categories_averages.csv))

Arts & Entertainment, Internet & Telecom, and News, Media & Publications have the highest total volume.

```{r volume_total_category_ordered, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font","volume_total_category_ordered.png"))
```

([data table](https://github.com/danielkupka/2020-google-searches/blob/master/proc_data/categories_total_volume.csv))

Finance has the highest mean cost per click.

```{r search_category_cpc_ordered, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "cpc_mean_category_ordered.png"))
```

([data table](https://github.com/danielkupka/2020-google-searches/blob/master/proc_data/categories_cpc.csv))

The average CPC for all searches is 0.61

<br> <br>

## 3.8 Keyword difficulty

Description of keyword difficulty from *ahref*:

> Keyword Difficulty (KD) effectively tells you the average height (backlink profile strength) of the other players in the game (top-ranking pages). The higher the number, the more height (backlinks) you'll likely to need to stand a chance of ranking.

As volume increases, the difficulty increases.

```{r difficulty_volume_sample, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_difficulty.png"))
```

The slope of the linear regression line is such that for each doubling of the volume, the difficulty increases by 1.63. For example, as the volume goes from 100 to 3200 (6 doublings), the difficulty increases by roughly 1.63 \* 6 \~= 10.

<br>

Higher difficulty also means higher CPC on average. Note that the Y axis is logarithmic, so a small move on the y axis is a large increase in value.

```{r difficulty_cpc_scatter, echo=FALSE}
knitr::include_graphics(here::here("plots", "0911", "difficulty_cpc_scatter.png"))
```

An alternative visualization of the same data by grouped category in a boxplot:

```{r difficulty_cpc_boxplot, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "cpc_difficulty_boxplot.png"))
```

<br> <br>

## 3.9 SERP features

Search Engine Results Pages (SERPs) are the pages that Google and other search engines show in response to a user's search query.

97.6% of the SERPs contain at least one SERP feature.

The figure below gives an overview of how often the different SERP features are shown. (Note there are (at least) two additional SERP feature types (*Knowledge Panel* and *Videos*), for which the sample size is too small to include.)

The SERP features shown in the most searches are *Image pack* and *People also ask*:

```{r serp_presence_repr, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "serp_features.png"))
```

The knowledge card has a huge effect in reducing the clicks-per-search, while the other SERP features have limited effect. Searches with the Shopping results SERP feature have higher cps on average.

```{r serp_cps, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "serp_cps.png"))
```

Easy keywords have fewer SERP features.

```{r serp_difficulty_boxplot, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "serp_difficulty_box.png"))
```

Thumbnail & Top stories is the most common SERP feature pairing.

```{r serp_pairs, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "serp_pairings.png"))
```

Searches without SERP features tend to be low volume.

```{r serp_volume_ridge, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "serp_volume.png"))
```

Searches with more SERP features have higher mean difficulty.

```{r serp_difficulty_n, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "serp_mean_difficulty.png"))
```

<br> <br>

## 3.10 Return rate

> Return Rate shows how often the same person searches for a given keyword over a 30-day period. A return rate of 1 means that people typically don't search for that keyword again within 30 days.

We arbitrarily grouped return rate into three groups, No significant return rate (\< 2), medium (2-10) and high (10+). Then we look at searches that have similar volume, but different return rate.

As the return rate increases, the number of clicks increases and the cost per click increases.

```{r return_rate, echo=FALSE}
title_description_table <- read.csv(here::here("plots", "csv", "return_rate.csv"), stringsAsFactors=F) %>% 
    rename("Return rate" = return_rate,
           "Mean return rate" = mean_return_rate,
           "Mean CPC" = mean_cpc,
           "Mean clicks" = mean_clicks,
           "Mean difficulty" = mean_difficulty)
knitr::kable(title_description_table, caption = "Comparison of searches with same volume but different return rates")
```

<br> <br>

## 3.11 International searches

There is search data from 5 English-speaking countries.

Of those, US and UK have the highest search volume per person. Americans use google 38% more than UK, 108% more than Canada and 90% more than Australia, when adjusted for population.

```{r countries_volume, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_population.png"))
```

US of course has much higher total search volume overall:

```{r countries_volume_prop, echo=FALSE}
knitr::include_graphics(here::here("plots", "1611", "search_volume_vs_us-1.png"))
```

US has significantly higher cost per click on average

```{r countries_cpc, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "cpc_countries.png"))
```

The following is based on analysis with *ahref*.

International searches have overall higher volume

```{r}
df <- tribble(~region, ~volume,
        "US", "33%",
        "International", "67%")
knitr::kable(df, caption = "Total search volume")
```

Internationally there are more searches with very low volume, while US has more searches with medium volume.

```{r comparison_count, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_region.png"))
```

There is not a large difference in the number of searches with very high volume. However, the total volume of these searches is a lot higher internationally

```{r comparison_volume, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_region_high.png"))
```

Searches that have high US volume tend to have high international volume, and vice versa. But there are some exceptions.

```{r international_v_us_volume, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_region_corr.png"))
```

*A version showing data points binned to hex tiles showing counts:*

```{r international_v_us_volume_hex, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_region_corr_hex.png"))
```

```{r tbl_int, echo=FALSE}
title_description_table <- read.csv(here::here("plots", "csv", "table_int.csv"), stringsAsFactors=F)
knitr::kable(title_description_table, caption = "Searches that have much higher volume internationally")
```

```{r tbl_us, echo=FALSE}
title_description_table <- read.csv(here::here("plots", "csv", "table_us.csv"), stringsAsFactors=F)
knitr::kable(title_description_table, caption = "Searches that have much higher volume in the US")
```

Searches that have higher volume in US have a higher click-per-search on average than searches that have higher volume internationally.

```{r international_cps, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_region_cps_corr.png"))
```

They also have a higher cost-per-click on average

```{r international_cpc, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_region_cpc_corr.png"))
```

Searches that have higher volume internationally, tend to have higher difficulty

```{r international_difficulty, echo=FALSE}
knitr::include_graphics(here::here("plots", "reworked_font", "volume_region_difficulty_corr.png"))
```

<br>

## 3.12 Searches with highest spikes in volume

Many of these are people featured in Google doodles

```{r tbl_spikes, echo=FALSE}
title_description_table <- vroom::vroom(here::here("plots", "csv", "spikes.csv"))
knitr::kable(title_description_table, caption = "Searches with highest difference in search volume over time")
```
